Set pose_pcd.hpp file
